% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTextSpan.R
\name{summaryTextSpan}
\alias{summaryTextSpan}
\title{Show a single value in a text span}
\usage{
summaryTextSpan(
  data,
  statistic = c("count", "first", "last", "min", "max", "mode", "concatenate", "unique"),
  column = NULL,
  selection = NULL,
  locale = "navigator.language",
  delim = ", ",
  sort = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data to summarize, normally an instance of \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{statistic}{The aggregation to perform.
Possible vales are \code{c("count","first", "last", "min","max","mode", "concatenate", "unique")}.
\code{first}, \code{last} and \code{concatenate} will select the value based on the index of the column passed to the function.
\code{min} and \code{max} will selected the value based on either d3.ascending (default) or through the function passed to the parameter \code{sort}
Count is the default.}

\item{column}{For most statistics, the column of \code{data} to summarize.
Not used for \code{"count"}.}

\item{selection}{Optional parameter. Expression to select a fixed subset of \code{data}. May be a logical vector or a one-sided formula that evaluates to a logical vector.
If used, the \code{key} given to \link[crosstalk:SharedData]{crosstalk::SharedData} must be a fixed column (not row numbers).}

\item{locale}{Optional parameter. For \code{count} and \code{unqiue} control how the number is displayed. Valid strings include
a \href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47 language tag},}

\item{delim}{Optional parameter. For concatenate, choose the joining string. Default is \code{", "}}

\item{sort}{Optional parameter. A sort function used to sort the data to determine the min or max value or the order in which to concatenate the value.}

\item{width}{Optional parameter. Fixed width for widget (in css units), which is passed to htmlwidgets::createWidget.}

\item{height}{Optional parameter. Fixed height for widget (in css units), which is passed to htmlwidgets::createWidget.}

\item{elementId}{Optional parameter. String used to define the element containing the widget, which is passed to htmlwidgets::createWidget.}
}
\description{
A \code{summaryTextSpan} displays a single value derived from a linked table.
Its primary use is with the \code{crosstalk} package. Used with \code{crosstalk},
a \code{summaryTextSpan} displays a value which updates as the data selection
changes.
}
