% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryValueBox.R
\name{summaryValueBox}
\alias{summaryValueBox}
\title{Create a value box component for a dashboard.}
\usage{
summaryValueBox(
  data,
  statistic = c("count", "sum", "mean", "pct_total", "sum_pct_total", "min", "max",
    "quantile"),
  column = NULL,
  selection = NULL,
  locale = "navigator.language",
  digits = 0,
  number_format = c("decimal", "percent", "currency", "unit"),
  signDisplay = c("auto", "always", "exceptZero", "negative", "never"),
  currency = NULL,
  unit = NULL,
  quantile = NULL,
  notation = c("standard", "scientific", "engineering", "compact"),
  caption = NULL,
  icon = NULL,
  color_thresholds = list(domain = c(0.5, 0.8), range = c("red", "gold",
    "mediumseagreen")),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data to summarize, normally an instance of \link[crosstalk:SharedData]{crosstalk::SharedData}.}

\item{statistic}{The statistic to compute.
Possible vales are \code{c("count", "sum", "mean", "pct_total", "sum_pct_total", "min", "max", "quantile")}.
Count is the default.}

\item{column}{For most statistics, the column of \code{data} to summarize.
Not used for \code{"count"} or \code{"pct_total"} statistic.}

\item{selection}{Expression to select a fixed subset of \code{data}. May be
a logical vector or a one-sided formula that evaluates to a logical vector.
If used, the \code{key} given to \link[crosstalk:SharedData]{crosstalk::SharedData} must be a fixed column (not row numbers).}

\item{locale}{Control how the number is displayed. Valid strings include
a \href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47 language tag},
or the string "navigator.language". Default value is "navigator.language"}

\item{digits}{Number of decimal places to display, or NULL to display full precision.
if number_format == \code{"percent"}, digits == \code{0} the statistics is 0.1234,
the resulting visualization will show "12\%".}

\item{number_format}{The resulting type of number to display.
Options are \code{c("decimal", "percent", "currency","unit")}. Decimal is the default.}

\item{signDisplay}{How to display the sign for the number. Possible values are:
\code{c("auto", "always","exceptZero","negative","never")}. Auto is the default}

\item{currency}{A string containing an ISO 4217 currency code. No default value.
\href{https://en.wikipedia.org/wiki/ISO_4217#List_of_ISO_4217_currency_codes}{Full list is availble here}
NOTE: If this value is not null, number_format will be reassigned to \code{"currency"}}

\item{unit}{A string containing a valid unit identifier from \href{https://tc39.es/ecma402/#table-sanctioned-single-unit-identifiers}{this list}
Pairs of simple units can be concatenated with "-per-" to make a compound unit.
NOTE: If this value is not null, number_format will be reassigned to \code{"unit"}}

\item{quantile}{A number in \verb{[0,1]}, specifying the quantile to use if statistic = "quantile".
The default number is 0.5 (i.e. median)}

\item{notation}{How the number should be displayed.
Possible vales are \code{c("standard", "scientific","engineering","compact")}.
Standard is the default.}

\item{caption}{Optiona parameter. The caption to display beneath the value.}

\item{icon}{Optional parameter. A string representing the name of a bootstrap icon.
See https://icons.getbootstrap.com/ for complete list of names}

\item{color_thresholds}{a list containing two named vectors, domain and range.
The domain is a series of numbers and range is a series of colour strings (hex or html colour names) such that for a value x:
\itemize{
\item if \code{x < domain[1]} then \code{colour == range[1]},
\item if \verb{domain[1] <= x < domain[2]} then \code{colour == range[2]}
\item ...
\item if \verb{domain[n-1] <= x < domain[n]} then \code{colour == range[n]}
\item if \code{domain[n] <= x} then \code{colour == range[n+1]}
}

As such, the length of the named vector range should be one item longer than the length of domain.
The default value for color_threshold is \verb{list(domain = c(0.5,0.8), range = c("red","gold","mediumseagreen")}).
To fix a single colour, \verb{let color_thresholds = list(domain = c(-Inf), range=c(unused_color,desired_color))}}

\item{width}{Optional parameter. Fixed width for widget (in css units), which is passed to htmlwidgets::createWidget.}

\item{height}{Optional parameter. Fixed height for widget (in css units), which is passed to htmlwidgets::createWidget.}

\item{elementId}{Optional parameter. String used to define the element containing the widget, which is passed to htmlwidgets::createWidget.}
}
\description{
A value box displays a value (usually a number) in large text, with a smaller
caption beneath, and a large icon on the right side.
}
\details{
See the flexdashboard website for additional documentation:
\url{https://pkgs.rstudio.com/flexdashboard/articles/using.html#value-boxes-1}
}
\examples{
library(flexdashboard)

valueBox(42, caption = "Errors", icon = "fa-thumbs-down")
valueBox(107, caption = "Trials", icon = "fa-tag")
valueBox(247, caption = "Connections", icon = "fa-random")

}
